Add hard-coded protocol interpreting support
This is a big commit, it adds in a protocol interpreting support, a
huge step towards the final goal. XML parser will be added in later
to replace the hard-coded ones.

This commit brings some problems that will be fixed later:
1. Output is fixed to interpret-mode (no original raw mode), I'll add
an option switch later
2. Compiler warnings because of the const stuff. I am too lazy to
change them. The problem will be eliminated with the XML parser which
dynamically generate the lists of protocol.
3. The hard-coded protocol is limited to core protocol and xdg_shell